+++ core/lib/merkle_tree/src/utils.rs
@@ -166,6 +166,49 @@ impl<T> Iterator for MergingIter<T> {
 impl<T> ExactSizeIterator for MergingIter<T> {}
 
 #[cfg(test)]
+pub(crate) mod testonly {
+    use crate::{Key, MerkleTree, PruneDatabase, TreeEntry, ValueHash};
+
+    pub(crate) fn setup_tree_with_stale_keys(db: impl PruneDatabase, incorrect_truncation: bool) {
+        let mut tree = MerkleTree::new(db).unwrap();
+        let kvs: Vec<_> = (0_u64..100)
+            .map(|i| TreeEntry::new(Key::from(i), i + 1, ValueHash::zero()))
+            .collect();
+        tree.extend(kvs).unwrap();
+
+        let overridden_kvs = vec![TreeEntry::new(
+            Key::from(0),
+            1,
+            ValueHash::repeat_byte(0xaa),
+        )];
+        tree.extend(overridden_kvs).unwrap();
+
+        let stale_keys = tree.db.stale_keys(1);
+        assert!(
+            stale_keys.iter().any(|key| !key.is_empty()),
+            "{stale_keys:?}"
+        );
+
+        // Revert `overridden_kvs`.
+        if incorrect_truncation {
+            tree.truncate_recent_versions_incorrectly(1).unwrap();
+        } else {
+            tree.truncate_recent_versions(1).unwrap();
+        }
+        assert_eq!(tree.latest_version(), Some(0));
+        let future_stale_keys = tree.db.stale_keys(1);
+        assert_eq!(future_stale_keys.is_empty(), !incorrect_truncation);
+
+        // Add a new version without the key. To make the matter more egregious, the inserted key
+        // differs from all existing keys, starting from the first nibble.
+        let new_key = Key::from_big_endian(&[0xaa; 32]);
+        let new_kvs = vec![TreeEntry::new(new_key, 101, ValueHash::repeat_byte(0xaa))];
+        tree.extend(new_kvs).unwrap();
+        assert_eq!(tree.latest_version(), Some(1));
+    }
+}
+
+#[cfg(test)]
 mod tests {
     use zksync_types::U256;
